src/python/requirements.txt
numpy
pillow
requests
tqdm
pytest
pandas
